新增建議售價功能

+++ src/main/java/tw/edu/ntub/imd/camping/service/UserService.java
@@ -14,7 +14,7 @@ public interface UserService extends BaseService<UserBean, String> {
 
     List<UserBadRecordBean> getBadRecord(String account);
 
-    int getComment(String account);
+    Double getComment(String account);
 
     void compensate(String account, CreditCard creditCard);
